EDNS
support EDNS (current version 0) with no options understood and the DO
flag left 0.

Actual maximum message size we wish to support should be a configuration
option in the server. We're also in error using 65535, as this ignores
the udp overhead. Something closer to, say, 8192 or so would be much
safer.

Removed the maximum message size as a construction parameter to the
connection. It doesn't belong in its per instance state. Instead add a
parameter to WriteTo().

The cache is already wasting space holding on to a complete record per
entry including the header. Now we've got more special case things to
put in the header, it is wasting even more.