Less() and Equal()
it isn't pleasant, but add Less() and Equal() to the RecordData
interface to implement conflict resolution for MDNS and equality
checking without reflect.DeepEqual()

brush up the resolver to differentiate being a recursive resolver vs.
asking recursive queries to a client.

add EmptyCache to allow a resolver client to run with no preloaded
assumptions.