update queries per RFC-2136
The server has outgrown being a single function handling StandardQuery.
Make it its own object, and move the logic of sending notifies there.

Fix the asymetry of Lable.Name() having to know how to escape zone file
format. Put that logic in TextCodec.putName() where it belongs.

Intern labels by name. This may or may not be faster, but to not do so
wastes a lot of memory. Also consider same pointer in Label.Equal()

Errors parsing names or labels from the wire are FormError. Keep two
non-protocol errors for parsing names or labels from text, and enforce
names can not be longer than 255 encoded bytes.

Close down established TCP connections with the server shutdown
context, and have dnsconn.Connection do any non-UDP writes with a 5
minute deadline.

Move the stupid ns.Zones.R out of that collection and in to the new
Server object where it belongs.