ACLs
define the ns.Access interface allowing the server to generically
control activity and access.

the daemon implements this with a simple ACL scanner. The REST server
applies this as well, although it is still not useful enough to do http
access for real.

recusive queries may be controlled in three different ways:
1- allow recursion, allow query of . (the cache)
2- do not allow recursion, but do allow query of . (the cache), which
will answer cached queries if it has them otherwise responding with
delegation
3- do not allow recusion or query of . (the cache), which will refuse
any queries outside the server's authority.