zone log files for updateable zones
updateable zones have their updates written to a log file. These patched
records are eventually folded back into the main files, preserving
interface specifics. If a zone is never updated, these files are never
touched and a log file never appears. A more desireable idea might be to
keep the log files themselves compacted rather than disturbing the
original files, although this presents other problems to the humans
editing these. A zone reload (or startup, which is a zone load) causes
this to occur.

add a type for the text codec to write and read for these log files.

add update functionality to dug, using the same log file syntax with the
annotations "prereq" and "update" to specify the sections of the update.

fix a bug in the wire codec allowing records with no rdata to be parsed.

fix a bug in the text codec to deal with lines containing only a comment
as a blank line. While I am not happy about having to look ahead in the
line reader to see if a line starts with a comment, the fact a blank at
the start of a line counts as a token makes this necessary. *sigh*

add general purpose null Access interface instances for always allow and
always deny usable from code.