record reform
RecordHeader was too big and too many things in too many places. Make
the record an interface like the data, and name the header and data 'H'
and 'D' rather than 'RecordHeader' and 'RecordData'. Handling of headers
is now a bit safer.

Still need to get the full headers out of the damn cache..