do not allocate 65535 size buffers every time
do not waste memory with the largest possible sized buffer for every
single operation. Break the free lists out so we have one of those still
for stream oriented connections, but have a smaller one for the udp
packets. Also allow this size to be administratively set so 8192 is just
a default. If this changes at runtime, the pool will adapt (at least
up). MDNS gets its own special pool with its well defined packet size.

Usage of EDNS is still a little clumsy.