determine interface of packets the modern way
Rather than try to be portable by binding to interface specific
addresses, determine the interface when the datagrams are received. This
works for just about everyone except Windows. Interface specific
listeners can still be created for Windows, just not automatically.

Doing it this way is necessary for multicast where, ironically, the
older bind to address method would only work on Windows.

This also means tcp connections cannot automatically do interface
specific queries, but that should not be a great limitation.