connection handling improvements
a StreamConn is really any connected net.Conn, of which packet based can
also be. Take advantage of this to allow a client to use a StreamConn
instead of a PacketConn if all they want to do is query a specific name
server or issue serial queries without the necessity or overhead of the
demux logic and associated go routine. This also makes it necessary to
add VC() to the Conn protocol, as doing a type assertion will not
necessarily tell if the connection is really packet based.

Do not send EDNS messages on stream oriented connections.

In the demux, if received packets are not claimed above a certain
backlog, prune them out.