test and connection cleanup
TestClose() was racing reading from a closed socket vs canceling a
context as it was accidentally using a closed connection. Reverse the
order of the tests, canceling with a context first, then closing.

sensibly regroup the functions together in connection.go keeping the
specific types together as they should be