db interface change
minor overhaul I've been wanting to do for a while. Leave the db
abstraction only worried about storing and retrieving of entire set of
records for the name, not the type and class. The amount of data per
name is small, and the implementation of the db is not the place to
worry about how to query, match, or update sets of records based on
their type and class.

This also makes the memory db's snapshotting quite trivial as the RRMap
elements are never modified, so a shallow copy of the map is all that is
needed to snapshot.