add ClientPort flag to Message
allow mDNS to determine a legacy unicast query by the fact the query did
not originate from 5353, and further allow us to ignore multicast answers.

If ClientPort is set in a received message, it means the source port was
not the same as the listening port. Only applies to PacketConn based
listeners, of course, and will never be set for non udp based protocols.