mDNS discovery
mDNS queries, an ipc mechanism, new 'dugs' cli

we completely ignore NSEC negative responses, big TODO is parsing them
to return NXDomain if asked for a name known not to have a certain
record type.

interface of answers is lost in the ipc data. This needs to be addressed
before we can support publishing.