mDNS interface info to ipc clients
plumb interface info to the ipc clients.

improve one shot queries.

fix a stupid bug I introduced to dnsconn.StreamConn with unix sockets.
They were not being treated as stream sockets.

nil vs. empty record decoding actually needs to stay as it was, but we
can make it mdns specific for the benefit of the ipc connection. Note,
however, mDNS can send empty records too, so this is not a good actual
solution; it would be far better to just return the NSEC records and let
the client take it apart.