mDNS reduce excessive queries
Do not always begin a persistent query with an initial one. Use the
cache if it is valid first, and then send an immediate query if found
not to be.

When requerying, do not use the TTL of the old records to determine next
requery, avoiding doubling up (or worse) the requery.