further persistent query improvements
do not combine the outgoing maintainance queries and results back to the
resolver in the same routine, for numerous reasons. Not only does it
make the code easier to follow, it greatly reduces the amount of
duplicate responses back to the resolver.

have the channel from Zone.PersistentQuery() fire only for the question
asked, not just the name, and fire only once per update. As we still
send n*m queries for protocol*interface, keep the silly debouncy logic
in place.