cleanup of stupid things
make Zone.Remove() less wtf-y by returning an IfaceRRSets for all
interfaces, which can then more easily be used for the mDNS routines it
is intended, as well as a lot more efficient.

Actually initialize the questions map so we don't crash when asked for
shared records we know about.

Shut down automatic host announcement in a way that doesn't potentially
cause it to think it conflicted when it stops, as well as restating code
that was just.. wrong.