self interrogation
a quick hack to have an empty question over ipc return the host
records so a client intending to publish services knows who we are.
Keeping a persistent query open on this is also helpful to be notified
if a conflict even occurs and to what the updated host name is, and its
status by having host records.