mdns IPC  robustness
opportunistically try to connect so client processes are usable if mDNS
server comes up later. If we fail a query for IPC connection reasons,
close the connection and set to nil for a retry.

avoid using Any for host queries as not all dns servers behave rationally
with it.

add hostname command to sd

improved TXT handling of services