exclusive record handling
The data model can in theory support a distinction between exclusive
ownership of a name vs. a name+type+class, but in practice trying to be
flexible enough to support this was stupid and the source of some
contradictory mDNS behavior. Put the flag in the name where it goes.

This also means we (finally) take care of the NSEC pseudo records
automatically.

Also fix a wild-ass wtf query format bug in probe.