limited chase CNAME from authority or cache
if we are answering from authority or cache, still do a limited CNAME
chase. If recusion is desired and we cannot satisfy the ultimate CNAME
destination, fall back to a new recursive query. (This means if a zone
has unresolved cnames within itself and a recursive query is performed,
the client will not get an authoritative answer, but such a situation is
also in error)