responsive one-shot queries
make a distinction between one-shot and persistent queries, allowing
one-shot queries to complete a lot faster. One-shot queries return
immediately if records are on any interface, even if they are missing
from others.

PersistentQuery logic in mdns.go is way more twisty to read than I'd
like it to be..