smarter NSEC records
instead of synthesizing nsec records during the query, make them when we
actually put the records in via announce. Not only does the amortize
creating them, they can be put in the non interface specific database so
queries for which no records exist on the interface may be quickly
answered. This greatly helps oneshot performance for exclusive entries
not appearing on all interfaces.