StreamConn peer info and verbose refused errors
A bit hard to create network based ACLs for TCP based queries when the
peer address is nil! Have a StreamConn return the peer and as a special
case, not complain about ErrIsConn if that's the same address being sent
to.

Also wrap dns.Refused responses in more verbose errors indicating why
they were refused.