delgation improvements
fix a bug recursively resolving ns records where the suffix only needs
to be taken if we got no answers from cache. That is, do not proceed
with the valid cache answers validating them as if they came from a
query for the suffix!

Let ResolveIP use TypeAny which will give available A or AAAA from the
cache. This is significant when chasing delegation as, for example, glue
A records come back in the additional, yet we waste time and resources
trying to fine the non-existent AAAA records instead of just using the A
records we have immediately. Or even if the AAAA records did exist, we
missed an opportunity to skip the intermediate query or queries needed
to find it. In simple terms, use the birds in hand rather than looking
for more in the bushes.