ask all nameservers..
when resolving any query with multiple name servers, stop on first
successful resolution, not first answer. This helps flaky situations
where one name server might refuse the query but the other has an
answer.

and therefore when building a list of nameservers to follow delegation,
resolve all of them to use (with luck, most of them will already be in
the cache via the additional section of the answer giving them to us)

todo: think about parallelizing the host queries for the ns records.